Removed unnecessary debug print statements and changed bullies to be a list instead of a dictionary.
Previously bullies was a dict with the bullies' twitter ids as keys and
each of their most recent tweets that have been analyzed by the bot as
the piece of data associated with that key. With the streaming api we no
longer need to keep track of their most recently made tweets. Keeping
the data in dict format doesn't make sense so I added a check to see if
the bullies list is a dict. If it is then it will change it to a list.

Hurrah for backwards compatability!